requirements.txt
streamlit==1.20.0
requests==2.25.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
python-dotenv==1.0.0